Initial implementation
The addrxlat API is mostly compatible with the code included in
libkdumpfile. The unit test was taken from the original project
and modified only slightly:

  * If a translation method has a full address of type NOADDR,
    this is returned as FullAddress(NOADDR, 0) rather than None.

  * Some exception types and messages are different.

OTOH the kdumpfile API compatibility has not been preserved. Most
importantly, a newly created kdumpfile.Context is not associated
with any underlying file. The dump file must be opened explicitly
using the open() method (or by assigning to the ATTR_FILE_FD
attribute).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>