Expose kdumpfile.read() and kdumpfile.read_string()
Oh no, these two crucial methods were forgotten.

The API is not compatible with the bindings that are built from the
libkdumpfile repository. In particular, read() returns a FFI buffer
object, so the [] operator returns bytes, not integers.

The read_string() also returns bytes, because a C string may not be valid
Unicode.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>