Fix use after free in Buffer objects
The Python constructor installs its own put_page callback. However, when
the Python Buffer object is created from a pointer to a pre-existing
addrxlat_buffer_t, the C struct lifetime is expected to be longer than that
of the Python object. The put_page callback may then be called after the
Python object has been already freed.

Save the original callback function and private data in the constructor and
restore them in the destructor.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>